+++ rust/src/reproducibility_spec/per_lang/mod.rs
@@ -1,3 +1,4 @@
+pub(super) mod apple;
 pub(super) mod cc;
 pub(super) mod container;
 pub(super) mod go;